chore: separate SheetOptions from SheetContent for the sake of the editor
